support Python 3
Add a little try/except block to support Python 3's re-naming of Queue to queue.

We try to import the modern module name first and fall back to Queue on
ImportError. This ordering is chosen a) if a modern queue module is provided as
a backport or monkey patch[1] this ordering will pick it up and b) trying to be
Python 3 "first" makes one feel all moral, etc :).

[1] cf the "future" library